requirements.txt
Flask
python-telegram-bot
openai
apscheduler
requests
python-dotenv
gunicorn
